% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmath.R
\name{crossprod}
\alias{crossprod}
\title{returns the cross product of 2 3D vectors}
\usage{
crossprod(x, y)
}
\arguments{
\item{x}{a vector of length 3}

\item{y}{a vector of length 3}
}
\value{
a vector of length 3
}
\description{
returns the cross product of 2 3D vectors
}
