% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmath.R
\name{dotprod}
\alias{dotprod}
\title{Returns the dot product of 2 3D vectors}
\usage{
dotprod(x, y)
}
\arguments{
\item{x}{a vector of length 3. ex. c(1,2,3)}

\item{y}{a vector of length 3}
}
\value{
The output is a number(scalar).
}
\description{
Returns the dot product of 2 3D vectors
}
