% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmath.R
\name{plot_crossprod}
\alias{plot_crossprod}
\title{Plots the cross product of 2 3D vectors, and the 2 inputted vectors on a 3D graph}
\usage{
plot_crossprod(x, y, unit = FALSE)
}
\arguments{
\item{x}{a vector of length 3.}

\item{y}{a vector of length 3.}

\item{unit}{default is set to FALSE. When set to TRUE, the function will plot all the vectors as unit vectors.
using TRUE will not give an accurate representation of the original vectors because they will all have a magnitude of 1.}
}
\value{
a 3D graph with the 2 original 3D vectors, and the cross product vector. This 3D graph can be rotated and zoomed in and out.
}
\description{
Plots the cross product of 2 3D vectors, and the 2 inputted vectors on a 3D graph
}
