% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmath.R
\name{unit_vector}
\alias{unit_vector}
\title{returns the unit vector of a 3D vector}
\usage{
unit_vector(x)
}
\arguments{
\item{x}{a vector of length 3}
}
\value{
a vector of length 3. This 3D vector will have a magnitude of 1
}
\description{
returns the unit vector of a 3D vector
}
