% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmath.R
\name{vector_mag}
\alias{vector_mag}
\title{finds the magnitude of a 3D vector}
\usage{
vector_mag(x)
}
\arguments{
\item{x}{a vector of length 3}
}
\value{
a number which is the magnitude of the vector
}
\description{
finds the magnitude of a 3D vector
}
